:recycle: api: move deleteSession usecase to src
