+++ api/src/certification/session/domain/usecases/delete-session.js
@@ -1,5 +1,14 @@
-import { SessionStartedDeletionError } from '../errors.js';
+import { SessionStartedDeletionError } from '../../../../../lib/domain/errors.js';
 
+/**
+ * @typedef {import ('../../../shared/domain/usecases/index.js').dependencies} deps
+ */
+
+/**
+ * @param {Object} params
+ * @param {deps['sessionRepository']} params.sessionRepository
+ * @param {deps['certificationCourseRepository']} params.certificationCourseRepository
+ */
 const deleteSession = async function ({ sessionId, sessionRepository, certificationCourseRepository }) {
   if (await _isSessionStarted(certificationCourseRepository, sessionId)) {
     throw new SessionStartedDeletionError();
